wrap taproot swap server calls in interface
