+++ libs/sdk-core/src/test_utils.rs
@@ -5,20 +5,31 @@ use std::{mem, vec};
 
 use anyhow::{Error, Result};
 use chrono::{SecondsFormat, Utc};
+use gl_client::bitcoin::blockdata::opcodes::all::{
+    OP_CHECKSIG, OP_CHECKSIGVERIFY, OP_CSV, OP_EQUALVERIFY, OP_HASH160,
+};
+use gl_client::bitcoin::blockdata::script;
+use gl_client::bitcoin::hashes::ripemd160;
+use gl_client::bitcoin::secp256k1::PublicKey;
+use gl_client::bitcoin::util::taproot::{TaprootBuilder, TaprootSpendInfo};
+use gl_client::bitcoin::{Address, Script, Sequence, XOnlyPublicKey};
 use gl_client::pb::cln::pay_response::PayStatus;
 use gl_client::pb::cln::Amount;
 use rand::distributions::uniform::{SampleRange, SampleUniform};
 use rand::distributions::{Alphanumeric, DistString, Standard};
 use rand::rngs::OsRng;
 use rand::{random, Rng};
-use sdk_common::grpc;
+use sdk_common::grpc::{
+    self, CreateSwapResponse, PaySwapResponse, RefundSwapResponse, SwapParameters,
+};
 use sdk_common::prelude::{FiatAPI, FiatCurrency, Rate};
+use secp256k1::musig::MusigKeyAggCache;
 use serde_json::{json, Value};
 use tokio::sync::{mpsc, watch, Mutex};
 use tokio::time::sleep;
 use tokio_stream::Stream;
 use tokio_stream::StreamExt;
-use tonic::Streaming;
+use tonic::{Status, Streaming};
 
 use crate::backup::{BackupState, BackupTransport};
 use crate::bitcoin::hashes::hex::ToHex;
@@ -39,6 +50,7 @@ use crate::models::{
     LspAPI, NodeState, Payment, ReverseSwapServiceAPI, SegwitSwapperAPI, SyncResponse, TlvEntry,
 };
 use crate::node_api::{CreateInvoiceRequest, FetchBolt11Result, NodeAPI, NodeError, NodeResult};
+use crate::swap_in_taproot::TaprootSwapperAPI;
 use crate::swap_out::boltzswap::{BoltzApiCreateReverseSwapResponse, BoltzApiReverseSwapStatus};
 use crate::swap_out::error::{ReverseSwapError, ReverseSwapResult};
 use crate::{
@@ -693,6 +705,105 @@ impl FiatAPI for MockBreezServer {
             value: 20_000.00,
         }])
     }
+}
+
+#[tonic::async_trait]
+impl TaprootSwapperAPI for MockBreezServer {
+    async fn create_swap(
+        &self,
+        hash: Vec<u8>,
+        refund_pubkey: Vec<u8>,
+    ) -> SdkResult<CreateSwapResponse> {
+        let lock_time = 1008;
+        let claim_pubkey_bytes =
+            hex::decode("0207121ffeda98fb0b28258d06efaa97623d1b4298dce269630432b9c12f4f6c84")
+                .unwrap();
+        let claim_pubkey = PublicKey::from_slice(&claim_pubkey_bytes).unwrap();
+        let refund_pubkey = PublicKey::from_slice(&refund_pubkey).unwrap();
+        let (x_only_claim_pubkey, _) = claim_pubkey.x_only_public_key();
+        let (x_only_refund_pubkey, _) = refund_pubkey.x_only_public_key();
+        let claim_script = claim_script(&x_only_claim_pubkey, &hash);
+        let refund_script = refund_script(&x_only_refund_pubkey, lock_time);
+
+        let taproot_spend_info = taproot_spend_info(
+            &claim_pubkey.serialize(),
+            &refund_pubkey.serialize(),
+            claim_script,
+            refund_script,
+        )?;
+        let address =
+            Address::p2tr_tweaked(taproot_spend_info.output_key(), Network::Bitcoin).to_string();
+
+        Ok(CreateSwapResponse {
+            address: address.to_string(),
+            claim_pubkey: claim_pubkey_bytes,
+            lock_time,
+            parameters: Some(SwapParameters::default()),
+        })
+    }
+    async fn pay_swap(&self, _payment_request: String) -> Result<PaySwapResponse, Status> {
+        Ok(PaySwapResponse::default())
+    }
+    async fn refund_swap(
+        &self,
+        _address: String,
+        _input_index: u32,
+        _pub_nonce: Vec<u8>,
+        _transaction: Vec<u8>,
+    ) -> SdkResult<RefundSwapResponse> {
+        Ok(RefundSwapResponse::default())
+    }
+    async fn swap_parameters(&self) -> SdkResult<Option<SwapParameters>> {
+        Ok(Some(SwapParameters::default()))
+    }
+}
+
+fn claim_script(x_only_claim_pubkey: &XOnlyPublicKey, hash: &[u8]) -> Script {
+    script::Builder::new()
+        .push_opcode(OP_HASH160)
+        .push_slice(&ripemd160::Hash::hash(hash))
+        .push_opcode(OP_EQUALVERIFY)
+        .push_x_only_key(x_only_claim_pubkey)
+        .push_opcode(OP_CHECKSIG)
+        .into_script()
+}
+
+fn refund_script(x_only_refund_pubkey: &XOnlyPublicKey, lock_time: u32) -> Script {
+    script::Builder::new()
+        .push_x_only_key(x_only_refund_pubkey)
+        .push_opcode(OP_CHECKSIGVERIFY)
+        .push_int(Sequence::from_height(lock_time as u16).to_consensus_u32() as i64)
+        .push_opcode(OP_CSV)
+        .into_script()
+}
+
+fn key_agg_cache(claim_pubkey: &[u8], refund_pubkey: &[u8]) -> Result<MusigKeyAggCache> {
+    let cp = secp256k1::PublicKey::from_slice(claim_pubkey)?;
+    let rp = secp256k1::PublicKey::from_slice(refund_pubkey)?;
+    Ok(MusigKeyAggCache::new(
+        &secp256k1::Secp256k1::new(),
+        &[&cp, &rp],
+    ))
+}
+
+fn taproot_spend_info(
+    claim_pubkey: &[u8],
+    refund_pubkey: &[u8],
+    claim_script: Script,
+    refund_script: Script,
+) -> Result<TaprootSpendInfo> {
+    let m = key_agg_cache(claim_pubkey, refund_pubkey)?;
+    let internal_key = m.agg_pk();
+
+    // Convert from one secp256k1 crate to the other.
+    let internal_key = XOnlyPublicKey::from_slice(&internal_key.serialize())?;
+
+    // claim and refund scripts go in a taptree.
+    Ok(TaprootBuilder::new()
+        .add_leaf(1, claim_script)?
+        .add_leaf(1, refund_script)?
+        .finalize(&Secp256k1::new(), internal_key)
+        .map_err(|_| anyhow::anyhow!("taproot builder error"))?)
 }
 
 pub struct MockBuyBitcoinService {}
